requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.1
numpy==2.2.2
openai==2.6.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
PyMySQL==1.1.2
pyparsing==3.2.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
requests==2.32.3
scikit-learn==1.7.2
scipy==1.16.2
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
